requirements.txt
django
psycopg2
dj_database_url
django-phonenumber-field
phonenumbers
Pillow